requirements/base.txt
# Общие зависимости
python-telegram-bot==21.0.1
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
redis==5.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
yookassa==3.2.0
google-genai==1.0.0
Pillow==10.2.0
aiohttp==3.9.1
